# openapi.yaml
openapi: 3.0.3
info:
  title: Flask CRUD
  version: 1.0.0
...
paths:
...
  /defense_mechanisms:
    post:
      summary: Create a defense mechanism
      tags:
        - DefenseMechanisms
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - mechanism
              properties:
                mechanism:
                  type: string
                  example: Firewall
      responses:
        "201":
          description: Defense mechanism created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefenseMechanism"
        "400":
          description: Invalid mechanism
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      summary: List all defense mechanisms
      tags:
        - DefenseMechanisms
      responses:
        "200":
          description: List of defense mechanisms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DefenseMechanism"
  /defense_mechanisms/{id}:
    get:
      summary: Get a defense mechanism by ID
      tags:
        - DefenseMechanisms
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: Defense mechanism found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefenseMechanism"
        "404":
          description: Defense mechanism not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Update a defense mechanism
      tags:
        - DefenseMechanisms
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - mechanism
              properties:
                mechanism:
                  type: string
                  example: IDS
      responses:
        "200":
          description: Defense mechanism updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefenseMechanism"
        "400":
          description: Invalid mechanism
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Defense mechanism not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete a defense mechanism
      tags:
        - DefenseMechanisms
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "204":
          description: Defense mechanism deleted
        "404":
          description: Defense mechanism not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Defense mechanism cannot be deleted due to foreign key constraint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /target_industries:
    post:
      summary: Create a target industry
      tags:
        - TargetIndustries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - industry
              properties:
                industry:
                  type: string
                  example: Finance
      responses:
        "201":
          description: Target industry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TargetIndustry"
        "400":
          description: Invalid industry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      summary: List all target industries
      tags:
        - TargetIndustries
      responses:
        "200":
          description: List of target industries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TargetIndustry"
  /target_industries/{id}:
    get:
      summary: Get a target industry by ID
      tags:
        - TargetIndustries
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: Target industry found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TargetIndustry"
        "404":
          description: Target industry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Update a target industry
      tags:
        - TargetIndustries
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - industry
              properties:
                industry:
                  type: string
                  example: Healthcare
      responses:
        "200":
          description: Target industry updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TargetIndustry"
        "400":
          description: Invalid industry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Target industry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete a target industry
      tags:
        - TargetIndustries
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "204":
          description: Target industry deleted
        "404":
          description: Target industry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Target industry cannot be deleted due to foreign key constraint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
...
    DefenseMechanism:
      type: object
      properties:
        id:
          type: integer
          example: 1
        mechanism:
          type: string
          example: Firewall
...
    TargetIndustry:
      type: object
      properties:
        id:
          type: integer
          example: 1
        industry:
          type: string
          example: Finance
    Error:
      type: object
      properties:
        error:
          type: string
          example: Not found